Create a symlink to cores directory based on core_pattern
Modifies the postStart hook to create a symlink to /mnt/disk0/cores at
the directory being used by kernel.core_pattern. This is useful when
the core_pattern's value points to some absolute path like
/var/lib/cores instead of a relative path and the path doesn't exist
inside the container.

Changing to bash as some of the bash constructs are being used. The
current container image's sh is just a symlink to bash.

Scenarios tested:
- Installed the chart on nodes having `kernel.core_pattern=/var/lib/test/cores123/core.%e.%p.%t`.
- Exec into the pod,
  ```console
  $ ls -ll /var/lib/test
  total 0
  lrwxrwxrwx 1 root root 16 Nov  2 13:45 cores123 -> /mnt/disk0/cores

  $ /home/yugabyte/dumpit
  Hello World!
  Segmentation fault (core dumped)

  $ pwd
  /mnt/disk0/cores

  $ ls
  core.dumpit.78.1604324807
  ```
- Installed the chart where core_pattern is set to pipe to some binary
  with `|/usr/lib/systemd/systemd-coredump %P %u %g %s %t %c %h %e`,
  as expected we don't do anything for such cases (can't to anything).

Ref: https://github.com/yugabyte/yugabyte-db/issues/4385#issuecomment-717656848

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>