[#7464] Charts: Make prometheus.yml writable by YW container
Summary: Platform HA needs the prometheus.yml to be writable by the YW container. This didn't work before since the current prometheus.yml is mounted from a configMap which is read-only. This change copies the configMap yaml to a yugaware-storage persistent volume + mounts this yaml in the prometheus container to use as well.

Test Plan:
Create a YW using the latest stable charts. Then upgrade it with these changes.

1. exec into the YW container and modify /prometheus_configs/prometheus.yml
2. exec into the prometheus container + cat /prometheus_configs/prometheus.yml to verify that the changes made in the YW container are reflected here too.

Reviewers: jvigil, arnav, rtsisyk, sanketh

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D10789