[yugabyte] Changes for multicluster solutions
Some of the multicluster solutions need this setup where we have one
service per pod, due to some bug or no support for Statefulset +
headless services on which we rely. multicluster.createServicePerPod
does that.

We have to use new naming style with a different release name across
multiple clusters, so that the Master and TServer pods get a unique
identity across clusters. In this case we still need a common name
across these clusters, multiple.createCommonTserverService creates a
service with a name independent of release name.

Scenarios tested:
- As this is behind a flag, this does not affect any existing
deployments.
- These flags have been tested with Istio multicluster setup.

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>