[Backport 2.7.1][#9297] Exec yugabyte server process
Summary:
This substitutes the bash process with the yugabyte server process.
Previously, the bash process captured any TERM signals received and did
not forward them to the yugabyte server child process. Therefore, the
the yugabyte server process would be shutdown uncleanly after the
shutdown grace period because it never received the shutdown signal. By
replacing the bash process with the yugabyte server (via exec), the
shutdown signal will be forwarded directly to the yugabyte server
process, allowing for a clean shutdown.

The original commit that introduced this problem was:
https://github.com/yugabyte/charts/commit/bcf1703873e0d79c62de4403c9d6f87009f99598

Test Plan:
Create yugabyte universe, manually delete a master or tserver pod. Verify that
the server process now receives a proper SIGTERM shutdown signal by checking
the logs for a message similar to:
```
*** Aborted at 1626270790 (unix time) try "date -d @1626270790" if you are using GNU date ***
PC: @                0x0 (unknown)
*** SIGTERM (@0x0) received by PID 1 (TID 0x7f71cf9042c0) from PID 0; stack trace: ***
    @     0x7f71bf5b1ba0 (unknown)
    @     0x7f71bf5b146d __GI___nanosleep
    @     0x7f71cb07f00c  yb::server::TotalMemWatcher::MemoryMonitoringLoop()
    @           0x4093b0  yb::master::MasterMain()
    @     0x7f71bec20825 __libc_start_main
    @           0x408849 _start
    @                0x0 (unknown)
```

Reviewers: sanketh

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D12273