Make PodDisruptionBudget configurable.
The old default of `maxUnavailable: 0` doesn't make sense because a StatefulSet doesn't allow extra replicas to be scheduled (not that kube tries to upsize collections anyways). This effectively means that the pods will never be allowed to be evicted which just results in inefficient use of a kube cluster and block updates. This makes the PDB configurable to allow testing of a better configuration.

Likely this will be updated to a different default value once it has been tested.