[PLAT-2036] Add resource requests to Platform helm chart
Summary:
Currently resource requests are not specified for postgres/prometheus containers, so the
Platform pod might be be placed on a node that doesn't have sufficient memory for prometheus. This
change allows specifying a resource request for postgres, prometehus, nginx containers so that the
pod can be sized appropriately.

Total requests for Yugaware now come to

memory: 9.3GB
cpu: 4.75

Test Plan:
helm install with the following command

```
 helm install --namespace test-sanketh-1 test-sanketh-1 yugaware --set=yugaware.multiTenant=true,image.repository=quay
.io/yugabyte/yugaware-itest,image.tag=2.9.2.0-b56,nginx.resources.requests.cpu=1,postgres.resources.requests.cpu=0.5,prometheus.resources.
requests.cpu=1
```

Verify that the overriden and default requests got applied

Reviewers: jvigil, nsingh

Reviewed By: jvigil, nsingh

Subscribers: alan, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D13564