Use policy/v1 of PodDisruptionBudget when it is available
The policy/v1beta1 of PodDisruptionBudget has been deprecated and will
be removed in Kubernetes version 1.25. With this change we use
policy/v1 of PodDisruptionBudget if it is available in the cluster or
else we fallback to policy/v1beta1 for both the charts yugabyte and
yugaware.

In case of `helm template` command, it will always report that the
policy/v1/PodDisruptionBudget is missing unless we provide the
`--api-versions policy/v1/PodDisruptionBudget` argument to it. The
only place where `helm template` is used, is the GKE marketplace's
deployer image, where don't have control over the exact `helm
template` command. In that case the pdbPolicyVersionOverride value of
yugaware chart can be used to explicitly set the APIVersion.

Scenarios tested:
1. Installed both charts on Kubernetes 1.20, and checked the
   APIVersion of PDB resources
   ```
   $ kubectl get pdb -n bg-dep \
       -ojsonpath='{range .items[*]}{.metadata.name}{" "}{.apiVersion}{"\n"}{end}'
     yb-master-pdb policy/v1beta1
     …
   ```

2. Upgraded the Kubernetes cluster to 1.21, and again checked the
   APIVersion of PDB resources with above command.
   ```
   yb-master-pdb policy/v1
   ```
3. Ran `helm upgrade` for both the chart, and checked the APIVersion
   again. The upgrade command was successful, and APIVersion remained
   policy/v1.

4. Tried the following template commands, which yield expected
   results.
   ```
   helm template ./yugabyte => uses v1beta1
   helm template ./yugaware => uses v1beta1
   helm template ./yugabyte --api-versions policy/v1/PodDisruptionBudget => uses v1
   helm template ./yugaware --api-versions policy/v1/PodDisruptionBudget => uses v1
   helm template ./yugaware --set pdbPolicyVersionOverride=v1 => uses v1
   ```

Fixes https://github.com/yugabyte/yugabyte-db/issues/10597

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>