[BACKPORT 2.8] Use policy/v1 of PodDisruptionBudget when available (#105)
Original commit: 6031111c0ec73cfc1512986f952992775f27eabe

The policy/v1beta1 of PodDisruptionBudget has been deprecated and will
be removed in Kubernetes version 1.25. With this change we use
policy/v1 of PodDisruptionBudget if it is available in the cluster or
else we fallback to policy/v1beta1 for both the charts yugabyte and
yugaware.

In case of `helm template` command, it will always report that the
policy/v1/PodDisruptionBudget is missing unless we provide the
`--api-versions policy/v1/PodDisruptionBudget` argument to it. The
only place where `helm template` is used, is the GKE marketplace's
deployer image, where don't have control over the exact `helm
template` command. In that case the pdbPolicyVersionOverride value of
yugaware chart can be used to explicitly set the APIVersion.

Scenarios tested:
1. Tried the following template commands, which yield expected
   results.
   ```
   helm template ./yugabyte => uses v1beta1
   helm template ./yugaware => uses v1beta1
   helm template ./yugabyte --api-versions policy/v1/PodDisruptionBudget => uses v1
   helm template ./yugaware --api-versions policy/v1/PodDisruptionBudget => uses v1
   helm template ./yugaware --set pdbPolicyVersionOverride=v1 => uses v1
   ```

Fixes https://github.com/yugabyte/yugabyte-db/issues/10597

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>