Add configmap to update main nginx configuration
Summary:
We might need to bump up the number of active connections to platform. This adds a
configmap to modify the main nginx so we can customize that as well as add more configurations later
down the line. It uses the default config generated by nginx and just adds the parameter to update
worker connections

Test Plan:
helm install platform. Verified that it comes up, and that the nginx.conf has the correct
number of worker connections set.

Reviewers: jvigil, kcox, daniel

Reviewed By: kcox, daniel

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D14696