[Plat-520] Added metric relabel for CPU metrics
Summary -

The kube-state-metric changed the following CPU metric name.
And the Yugabyte Platform java code still uses the old metric name to show the CPU metric on Platform UI.
At this place, we had two ways to solve this. Either update the java code or handle it in the Prometheus scrape configuration.
We have chosen the second approach and fixed the Prometheus scrape configuration.

Test -
Deployed the platform using without this change and created the universe.
We didn't have CPU metrics in the Platform UI at this time.
Once I modified the Prometheus config using the helm upgrade, it started showing the CPU metrics.