[PLAT-5217] YBC changes to add releases volume mount
Summary: The ybc releases are being released from the Platform currently. The ybc build is packed in the yugaware build and later moved to the appropriate Platform directory. From this directory location, the ybc releases are managed by the Platform. As part of this release management strategy we have, we would like to mount this location to the Platform. The changes made in this PR are for this purpose.

Test Plan: Verify the volume mount

Reviewers: sanketh, sneelakantan

Reviewed By: sneelakantan

Subscribers: vpatibandla

Differential Revision: https://phabricator.dev.yugabyte.com/D19193