[PLAT-4338] Enable https on master/tserver web UI when node-to-node encryption is enabled
Summary:
Related to https://phabricator.dev.yugabyte.com/D19465

Allow newly created universes with db versions at least 2.17.1.0-b14 to have https enabled on master/tserver web UI. We will piggyback node-to-node encryption for this first phase. That is, on universe creation, if universe db version is at least 2.17.1.0-b14 and node-to-node encryption is set as true, we will enable https, otherwise, we will set to http as before.

To keep track of this, we have a `httpsEnabledUI` key in the universe config which has the values `false|true`, which determines whether we use http/https respectively. If `httpsEnabledUI` doesn't exist, we assume that the value is false.

Already created universes with db version at least 2.17.1.0-b14 will currently not have their https unless they disable and then enable node-to-node encryption.

We also have a new runtime config, called `node_ui` which will current be used to create a custom wsClient, which allows you to make api requests, is configured to currently accept any certificates and has a specified timeout.

From this change: https://github.com/yugabyte/yugabyte-db/commit/b737017d49a1257146f7af5b663d4edf8fe12ceb. We currently able to just use our http endpoint when making requests. The request will be redirected to the associated `https` endpoint as required.

Related to the above change, to enable https, we need the follow gflags set:
1. `webserver_redirect_http_to_https` - Allow node UI endpoints to be redirected from http -> https automatically
2. `webserver_certificate_file`  - certificate file
3. `webserver_private_key_file ` - private key
4. `webserver_ca_certificate_file` - the certificate's CA

For this commit, if the httpsEnabledUI bit is bi-directional.

However, we currently don't support turning https on if the universe is upgraded to 2.17.1.0-b14. However, a user can turn off then turn on node-to-node encryption to enable this

Test Plan:
Fixed unit tests

Made sure that all locations where node endpoints are used works when we set the gflags. This includes:
1. UniverseCRUDHandler (createUniverse) - check that we set the config properly
2. SessionController (proxyRequest) - When in the nodes page, we are able to click the master/tserver urls and whether or not the current node is a master leader, we should see the entire UI show the correct information.
3. TabletServerController (listTabletServers) - In the nodes page, we should be able to see the RAM used, SST size, etc.
4. prometheus.yml - Go to localhost:9090 and make sure that when we make a query on a node in the universe, we should get results back. Ex: `max_over_time(up{node_prefix="yb-admin-cwang-21710b19-universe"}[30s])`. Can also go to metrics page and check that the graphs are showing properly.
5. node_health.py (http_request) - Make sure that when going to the health tab, all checks should pass.
6. LiveQueryExecutor (call) - run yb-sample-apps, Ex: `docker run -d yugabytedb/yb-sample-apps --workload SqlInserts --nodes 10.9.119.167:5433,10.9.135.35:5433,10.9.204.146:5433 --username yugabyte --password <password>`. In the queries page, go to live queries and you should be able to see inserts into the table.
7. QueryHelper (queryUniverseNodes)
8. TlsToggle - Check that universes with version >= 2.17.1.0-b14 can have https turned on/off based on whether node-to-node encryption is turned on/off

Reviewers: sanketh, sb-yb

Reviewed By: sanketh, sb-yb

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D19468