[PLAT-7075] Chart changes needed to deploy ybc container on the pods
Summary: Make the required changes to run YBC as a sidecar container on tserver pods. These chart changes would bring up the YBC container on the tserver and expose the service on port 18018.

Test Plan:
Created an RF 3 universe with 3 master and 3 tserver pods.
Verified that YBC is running on all the tserver pods.
Took a backup on this universe.
Performed a restore operation on the universe using the YBC backup

Reviewers: sanketh, sneelakantan, bgandhi

Reviewed By: sneelakantan, bgandhi

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D23034