[Backport 2.17.2] [PLAT-6864] Easy Deployment of K8s cluster, changes to rbac rules.
Summary:
We want to simplify the deployment of K8s provider for yugaware instances
that are deployed on k8s. This needs an expansion of rbac rules.
This change merges in the rules from yugabyte db helm chart.

Test Plan:
Deployed and tested manually verified deployment went to helm chart deployment stage.

# Tip: Write "Fixes T123" in your summary to automatically close the
# corresponding task when this change lands.

# Tip: You can add a project as a subscriber or reviewer by writing
# "#projectname" in the appropriate field.

Reviewers: nbhatia, sanketh, sneelakantan

Reviewed By: sneelakantan

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D23389