[PLAT-7742] Fix backup and restore on auth-enabled k8s universes
Summary:
Create a volume which is empty and attach it as volumeMount /tmp to the tserver and yb-controller container.

Test Plan:
Created an auth-enabled universe with the above change. Verified that the universe creation is successful.
Took a backup on the above universe and performed a restore operation.
Created a file in the /tmp folder. Did a rolling restart and verified that this file does not exist.

Reviewers: sanketh, sneelakantan, anijhawan

Reviewed By: anijhawan

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D23683