[PLAT-5268][yugabyte] Fix YCQL for Istio MCS enabled clusters
Summary:
We bind to 0.0.0.0 for Istio as it requires the application to be
listening on 127.0.0.1. Due to
https://github.com/yugabyte/yugabyte-db/issues/16155 YCQL fail as it
tries to resolve internal FQDN of the pods from other clusters. With
this change we set the rpc_bind_address to POD_IP,127.0.0.1 so that
YCQL uses the POD_IP for its work, and Istio gets DB process listening
on localhost. We are not changing anything for Master in case of
Istio, it still listens on 0.0.0.0.

This change adds an extra blank line after each preflight check, I was
not able to get rid of it without breaking other checks. It doesn't
have any impact on the functionality.

Test Plan:
Applied these changes to a release Helm tar and deployed universe
using YBA in Istio MCS setup. I was able to create tables, run SELECT
query on those for both YCQL and YSQL.

The listen_address is the POD_IP, and broadcast_address/rpc_address is
the service IP.

```
$ select * from system.local;

 key   | bootstrapped | broadcast_address | cluster_name  | cql_version | data_center | gossip_generation | host_id                              | listen_address | native_protocol_version | partitioner                                 | rack          | release_version | rpc_address  | schema_version                       | thrift_version | tokens | truncated_at
-------+--------------+-------------------+---------------+-------------+-------------+-------------------+--------------------------------------+----------------+-------------------------+---------------------------------------------+---------------+-----------------+--------------+--------------------------------------+----------------+--------+--------------
 local |    COMPLETED |      10.64.11.135 | local cluster |       3.4.2 |  south-asia |                 0 | 3a6cb198-3f63-1c88-7544-aa5e5d5afe19 |      10.60.5.4 |                       4 | org.apache.cassandra.dht.Murmur3Partitioner | asia-south2-a |    3.9-SNAPSHOT | 10.64.11.135 | 00000000-0000-0000-0000-000000000000 |         20.1.0 |  {'0'} |         null
```

The peer and rpc_address are the service IPs, and preferred_ip is the POD_IP.
```
$ select * from system.peers;

 peer        | data_center | host_id                              | preferred_ip | rack          | release_version | rpc_address | schema_version                       | tokens
-------------+-------------+--------------------------------------+--------------+---------------+-----------------+-------------+--------------------------------------+--------------------------
 10.4.18.136 |  ap-south-1 | f374c786-9519-63ab-7c49-082b6776704b |    10.8.4.90 | asia-south1-b |    3.9-SNAPSHOT | 10.4.18.136 | 00000000-0000-0000-0000-000000000000 |  {'6148820866244280320'}
 10.4.22.239 |  ap-south-1 | 493c59ed-2d4b-f284-eb42-1118f901d392 |     10.8.0.4 | asia-south1-a |    3.9-SNAPSHOT | 10.4.22.239 | 00000000-0000-0000-0000-000000000000 | {'-6149102341220990976'}
```

Executed following helm template commands to ensure the generated
YAMLs have appropriate rpc_bind_address and correct preflight checks.

```
$ helm template . --set multicluster.createServiceExports=true --set multicluster.kubernetesClusterId=c1
$ helm template . --set istioCompatibility.enabled=true
$ helm template . --set istioCompatibility.enabled=true --set multicluster.createServicePerPod=true
$ helm template .
```

Reviewers: sanketh

Reviewed By: sanketh

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D23968