[PLAT-7765] Health checks failing for ybc on k8s multi-cluster deployments. Modify the dns1 format
Summary:
Defensive approach to delete the yb-controller.pid file if it exists during pod restart. This should ideally never happen. Added it just as a guardrail
Modify the DNS format of $dns1 to have a single wildcard
Added $node to the SAN list

Test Plan:
dev itests
Created a multi-cluster universe with these chart changes and verified that the health checks are succeeding.

Reviewers: bgandhi, amahajan, sanketh, anijhawan, sneelakantan

Reviewed By: anijhawan, sneelakantan

Subscribers: jenkins-bot, anijhawan

Differential Revision: https://phabricator.dev.yugabyte.com/D24208