[BACKPORT 2.17.3][PLAT-8514] Helm install with user provided key and certificate.
Summary:
When user provides base64 encoded key and certificate, with commit
6ddf0f0b4c30731dd498db6fd73aa8e2ccb533b3 helm install failed. With
this commit server.pem is a concatenation of key+cert. Hence, the
the correct way is to store the base64 encoded value of key+cert
together.

Test Plan: Manual helm install with key and certificate provided in values yaml.

Reviewers: anijhawan

Reviewed By: anijhawan

Subscribers: jenkins-bot

Differential Revision: https://phabricator.dev.yugabyte.com/D24685