[PLAT-8544] Merge affinity rules correctly
Affinity rules were not merging correctly, allowing customers to fully
overwrite nodeAffinity when they meant to add to it.

We now merge on both nodeAffinity and podAntiAffinity correctly. If
a user desires, they are still able to overwrite fields within those, but
it is not possible to add more affinity rules without overwriting the
existing rules