[PLAT-8544] Merge affinity rules correctly
Summary:
Affinity rules were not merging correctly, allowing customers to fully
overwrite nodeAffinity when they meant to add to it.

We now merge on both nodeAffinity and podAntiAffinity correctly. If
a user desires, they are still able to overwrite fields within those, but
it is not possible to add more affinity rules without overwriting the
existing rules

Test Plan:
tested with tserver and master custom affinities of podAntiAffinity and
nodeAffinity, and validated they merged correctly.
helm template .

Re-ran the same test with
helm template . --set AZ=test

to validate the workflow with and without az set

Reviewers: bgandhi, sanketh, sneelakantan, anijhawan

Reviewed By: bgandhi, anijhawan

Subscribers: jenkins-bot, anijhawan

Differential Revision: https://phorge.dev.yugabyte.com/D25142