Logical revert of the earlier revert with one change. Moved CRD out of the yugaware chart.revert
Summary:
Helm doesn't handle CRD's very well.
1. It needs CRDs to be in a special folder under template called crd.
2. Helm doesn't upgrade CRD in place and needs a special flag --skip-crd otherwise it fails the helm install.
https://helm.sh/docs/chart_best_practices/custom_resource_definitions/#install-a-crd-declaration-before-using-the-resource
In light of these, we reverted the earlier change and are changing our decision to bundle CRD with the yugaware chart itself.

However we still need a place to store the CRD that is versioned along with the operator/yba version hence we are currently storing it at the top level in the charts repo.

Test Plan: itest

Reviewers: dshubin, sanketh, bgandhi, anabaria

Reviewed By: dshubin, bgandhi

Differential Revision: https://phorge.dev.yugabyte.com/D26032