[Backport 2.19.2] [PLAT-10211]  Fixing Permissions for YBA upgrade to work as expected on k8s with files created with older permissions
Summary:
Previous version of YBA was creating  before the securityContext changes was  creating the files as owned by root,
post upgrade when yugaware container runs as yugabyte (10001), these files are still owned by root as they get
created in the volume /opt/yugabyte/yugaware/data/provision .
When Migration runs to recreate provision script this fails with a permission issue during the chmod portion
because we don't have permissions in yugaware container to do the chmod.
The fix is to do an explicit permission change for everything in the mounted volume before yugaware starts in an initContainer.
We already do something similar for prometheus files in the initContainer
we will have to extend that to modify ownership of everything in /opt/yugabyte/yugaware/data/provision

Test Plan:
Create a YBA with securityContext disabled, upgrade to YBA with secuityContext enabled
Expected result: init-permissions code should execute and yba should come up, files should be owned by yugabyte

Reviewers: dshubin, vkumar, amalyshev, sanketh, bgandhi

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D28316