[PLAT-9486][yugabyte] Write to files & run sync in liveness probe
Summary:
Instead of just runing the touch which might not reliably surface disk
failures, we now write current timestamp to the files on disk and then
run sync to ensure the write goes to the disk.

As part of the commit the commands are extracted as a template and
used liveness probe as well as before we start the DB process. In case
of write failures tee command returns non-zero exit code.

Test Plan:
- Installed the Helm chart with these changes and verified the content
  of the disk.check file.
- The files had timestamps which keep changing when the liveness probe
  runs.

```
[root@yb-master-0 disk0]# cat /mnt/disk1/disk.check
disk check at: Wed Sep 13 10:59:06 UTC 2023
[root@yb-master-0 disk0]# cat /mnt/disk0/disk.check
disk check at: Wed Sep 13 10:59:06 UTC 2023
```

Reviewers: anijhawan, sanketh, sneelakantan

Reviewed By: anijhawan, sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D28496