[PLAT-10477] helm option to create first customer in yba
Summary:
all yba workflows require a customer. Since we don't want users to use the UI
when using the operator (and, OSS users won't have a UI), we need
a way for the first "customer" to be created outside of the normal
register workflow.

This allows users to specify basic customer info via values.yaml

Test Plan: ensured a customer got created

Reviewers: anijhawan

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D30083