[PLAT-9097][yba] Add option to disable scraping of K8s nodes
Summary:
There are some users who don't manage Kubernetes based universes from
YBA deployed on Kubernetes. In that case they can disable the scraping
of Kubernetes nodes we do for a few metrics to show cpu, memory and
disk utilization graphs in the UI for Kubernetes based universes.

Adds a new flag under prometheus which can be set to false to disable
this scraping.

Test Plan:
- Deployed YBA with this change. Was able to see the graphs with
  default settings.
- Set scrapeKubernetesNodes to false. And the metrics were not
  scraped. Verified by going to Prometheus configuration.

Reviewers: vkumar, anijhawan, sneelakantan, sanketh

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D30447