[PLAT-17026] Add pull secret, node selector, affinity rules to the job thats created for customer creation for operator
Summary:
[PLAT-17026] Add pull secret, node selector, affinity rules to the job thats created for customer creation for operator
This is to make sure that the job runs on the same node selector parameters as the yugaware pod.
And it uses the same pull secret as the yugaware pod.

Test Plan:
Did helm template verified job was getting generated.
installed helm chart with non existant node selector, verified pod, job both remained in pending state.
verified pull secret was populated correctly on the job.

Reviewers: dshubin, vkumar, vbansal

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D42505