[#26624] yugabyted: Add environment variable to yugabyted-ui container
Summary:
Adds an environemnt variable to the yugabyted-ui container to allow for the UI to determine if it is in a k8s deployment.
This is useful for cases where we want the UI to behave differently for k8s deployments, such as in D43181.

Test Plan: no test plan

Reviewers: nikhil

Reviewed By: nikhil

Subscribers: yugabyted-dev, djiang, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43182