[BACKPORT 2024.2][PLAT-17513] Fix istio failing with tls enabled universes
Summary:
ORIGINAL DIFF: 8a05a551d46f03e13523dc9f136641cf360efd4f / D43740
Istio sidecar does some logic handling for grpc named ports. But when tls is enabled in the universe
it is not able to decipher the traffic and fails. Because of this internal pod ping works but pod to
pod and yba to pod fails. We should name the ybc port also as "tcp-".

Ref: https://istio.io/latest/docs/ops/configuration/traffic-management/protocol-selection/#explicit-protocol-selection

Test Plan: Tested the change with istio and tls enabled universe

Reviewers: sneelakantan, anabaria

Reviewed By: anabaria

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43843