[BACKPORT 2024.2][PLAT-17513] Fix istio failing with tls enabled universes
Summary:
Original commit: 2f391f3621b1fe8ddf64164acc12cfa2aa3f082f / D43885
Istio sidecar does some logic handling for grpc named ports. But when tls is enabled in the universe
it is not able to decipher the traffic and fails. Because of this internal pod ping works but pod to
pod and yba to pod fails. We should name the ybc port also as "tcp-".

The helm logic will now check for both names in service ports and assign the name "tcp-ybc-port" in container
port and services to the ybc port.

Ref:
https://istio.io/latest/docs/ops/configuration/traffic-management/protocol-selection/#explicit-protocol-selection

Test Plan: dev itests

Reviewers: sneelakantan, anabaria

Reviewed By: anabaria

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43940