[BACKPORT-2025.1][GH-#200]Make cert manager script compatiable with non-root user
Summary: The initial boot logic for our containers attempted to create a directory within /opt. Since /opt is generally owned by root, this action resulted in a failure of the universe creation task when the universe was created with a non-root security context. To resolve this, we are now mounting /opt/certs as an empty directory instead of programmatically creating it.

Test Plan:
Tested manually by -

  - Creating a universe with cert-manager + non-root security context
  - Creating a cert-manager only universe (regression)

Reviewers: vkumar, sneelakantan

Reviewed By: vkumar

Subscribers: yugaware, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D46298