[BACKPORT 2.25.2][#26624] yugabyted: Add environment variable to yugabyted-ui container
Summary:
Adds an environemnt variable to the yugabyted-ui container to allow for the UI to determine if it is in a k8s deployment.
This is useful for cases where we want the UI to behave differently for k8s deployments, such as in D43181.

Original commit: 15c45fd9513321809b9537407af6fbcac19ec222 / D43182

Test Plan: no test plan

Reviewers: nikhil

Reviewed By: nikhil

Subscribers: yugaware, djiang, yugabyted-dev

Differential Revision: https://phorge.dev.yugabyte.com/D43896