[PLAT-11861] Setting default resources for yb controller and yb cleanup container
Summary:
ybc has 1 cpu and 1gb memory defaults
yb-cleanup has 1/4 of that

Test Plan:
Validated that the resources are correctly applied to pods on deployment.

In addition, ran a backup and saw that YBC typically uses less the 1 core and 1 GB of memory,
so these defaults should be sane for the majority of deployments.

Reviewers: vkumar, anijhawan, sneelakantan, vbansal

Reviewed By: vkumar, vbansal

Subscribers: vbansal, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D44113