[PLAT-18055][PLAT-18054][PLAT-18045] Add option to provide ipFamilies/ipFamilyPolicy to services and set arbitrary args to serviceEndpoints
Summary:
We were seeing issues with customers not able to use dual stack cluster as the headless
service was not correctly setting preference for IPv6 and IPv4 addresses. Fixed by adding explicit
option to configure that.

Also added support to configure extraArgs and set arbitrary values to serviceEndpoints as it is easy
to miss some fields which are configurable in services.

Another change is conditionally setting `externalTrafficPolicy` as that is not a valid setting for
ClusterIP based service. Only non-ClusterIP services will be setting this field now.

Test Plan:
- Verified universe comes up with clusterIP based serviceEndpoint
- Verified extraArgs for service works as expected
- Verified DualStack with IPv6 preference

```
ipFamilies: ["IPv6", "IPv4"]
ipFamilyPolicy: "PreferDualStack"

serviceEndpoints:
  - name: "yb-tserver-service"
    type: ClusterIP
    scope: "Namespaced"
    app: "yb-tserver"
    ports:
      tcp-yql-port: "9042"
      tcp-yedis-port: "6379"
      tcp-ysql-port: "5433"
    extraArgs:
      internalTrafficPolicy: Cluster
```

Reviewers: anijhawan, anabaria, sneelakantan

Reviewed By: anijhawan, anabaria

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D45233