[BACKPORT 2025.1.0][PLAT-18177] Add use node hostname for master flags
Summary:
In [PLAT-18177] we saw that on kubernetes universes when node to node encryption is turned on pg upgrade breaks with the error message
"""
2025-07-25 14:13:26.385 UTC [896] FATAL: Handshake failed: Network error (yb/rpc/secure_stream.cc:1191): Endpoint does not match, address: 10.0.2.227, hostname: 10.0.2.227
"""
The code that causes this error is https://github.com/yugabyte/yugabyte-db/blob/88880a4a6875e440dc537d089a28294ac0b7e420/src/yb/rpc/secure_stream.cc#L1190

This error is caused if the server name mentioned on the cert and the hostname of postgres server does not match.
To fix this we set the use_node_hostname_for_local_tserver=true in master so that the postgres process started during pg_upgrade can get the same flags as well.

Test Plan: itest

Reviewers: vkumar

Reviewed By: vkumar

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D45615