[PLAT-18438] Set publishNotReadyAddress in all cases for headless service
Summary:
It is beneficial to set `publishNotReadyAddress` for all cases. This makes
sure we're atleast letting the DNS records be available for the pods which are
marked not-ready. For instance if a non-app container dies - like YBC, the peers
should continue to resolve address, even if the pod is not marked ready.

Test Plan: dev itests

Reviewers: anijhawan, sanketh

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46381