[PLAT-18453]Add YBC gflags checksum annotation to tserver pod spec
Summary:
Added YBC gflags checksum to tserver pod spec so that restarts can be triggered
when required for YBC gflags changes.
This is the same behavior as tserver/master gflags.

This is not widely used as the gflags rotation API for YBC is internal and YBC gflags
apart from throttle params don't need to be modified in most cases.

Test Plan: Verified gflags checksum is applied

Reviewers: anijhawan, dshubin

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D46443